Lab 2 - Training a ML box detector/requirements.txt
accelerate==0.18.0
aiofiles==23.1.0
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
aiosqlite==0.18.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-cognitiveservices-vision-customvision==3.1.0
azure-cognitiveservices-vision-face==0.6.0
azure-common==1.1.28
azure-core==1.26.1
azure-storage-blob==12.16.0
azure-storage-queue==12.6.0
Babel==2.12.1
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.0
billiard==3.6.4.0
bitsandbytes==0.38.1
bleach==6.0.0
blinker==1.5
cached-property==1.5.2
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
datasets==2.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.6
Django==4.1.4
django-widget-tweaks==1.4.12
dominate==2.7.0
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.8.2
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-Mail==0.9.1
Flask-WTF==1.1.1
flax==0.3.3
fonttools==4.39.2
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.4.0
google-api-core==2.11.0
google-api-python-client==2.69.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.57.0
httplib2==0.21.0
huggingface-hub==0.13.4
idna==3.4
iniconfig==2.0.0
ipykernel==6.22.0
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.4.8
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-server==1.23.6
jupyter-ydoc==0.2.3
jupyter_client==7.4.1
jupyter_core==5.3.0
jupyter_server_fileid==0.8.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
jupyterlab_server==2.21.0
kiwisolver==1.4.4
loguru==0.6.0
loralib==0.1.1
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
ml-dtypes==0.1.0
mpmath==1.2.1
msgpack==1.0.5
msrest==0.7.1
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==0.4.3
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.0
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.10.1
peft @ git+https://github.com/huggingface/peft.git@b21559e042555a8370f4b3cf1ad581c45291cf15
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.5.2
pluggy==1.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.31
protobuf==4.21.11
psutil==5.9.4
psycopg2==2.9.3
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.0
PyDrive==1.3.1
Pygments==2.14.0
PyJWT==2.7.0
pyparsing==3.0.9
pypiwin32==223
pyrsistent==0.19.3
pytest==7.3.0
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2022.2.1
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
qrcode==7.3.1
qtconsole==5.4.2
QtPy==2.3.1
regex==2023.3.23
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
Send2Trash==1.8.0
sentencepiece==0.1.98
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
sqlparse==0.4.2
stack-data==0.6.2
stripe==5.0.0
sympy==1.11.1
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.0+cu117
torchaudio==2.0.1+cu117
torchvision==0.15.1+cu117
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers @ git+https://github.com/huggingface/transformers@28f26c107b4a1c5c7e32ed4d9575622da0627a40
twilio==8.2.0
typing_extensions==4.3.0
tzdata==2022.2
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.12
uttlv==0.6.0
vine==5.0.0
virtualenv==20.17.1
visitor==0.1.3
voila==0.4.0
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
websockets==10.4
Werkzeug==2.2.2
whitenoise==6.1.0
widgetsnbextension==4.0.7
win32-setctime==1.1.0
WTForms==3.0.1
xxhash==3.2.0
y-py==0.5.9
yarl==1.8.2
ypy-websocket==0.8.2
